pyproject.toml
[project]
name = "kip"
version = "0.1.0"
description = ""
authors = [
    {name = "riekehof",email = "riekehof@iabg.de"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "sounddevice (==0.5.3)",
    "numpy (==2.3.5)",
    "faster-whisper (==1.2.1)",
    "pyannote-audio (==4.0.2)",
    "torch (==2.8.0)",
    "reportlab (>=4.4.5,<5.0.0)"
]
